requirements.txt
google-genai
streamlit
ddgs
fpdf2
chromadb
pymupdf
